Update 25 plugins: Advanced Woo Search, Backup, Contact Form 7, Dokan, Elementor, Enhanced E-commerce for Woocommerce store, FooGallery, Google Analytics for WordPress by MonsterInsights, Gutenberg, Ninja Forms, PayPal Donation, Regenerate Thumbnails, Social Reviews & Recommendations, WIP Custom Login, WooCommerce Admin, WIP WooCarousel Lite, WooCommerce PayPal Checkout Gateway, WooCommerce Services, WP RSS Aggregator, YITH WooCommerce Product Bundles, YITH WooCommerce Added to Cart Popup

+++ wp-content/plugins/backup/com/core/SGBoot.php
@@ -266,7 +266,7 @@ class SGBoot
 
 	public static function isFeatureAvailable($feature)
 	{
-		return (SGConfig::get('SG_FEATURE_'.strtoupper($feature))===1?true:false);
+		return ((int)SGConfig::get('SG_FEATURE_'.strtoupper($feature))===1?true:false);
 	}
 
 	private static function prepare()
